+++ src/app/beer_garden/db/mongo/util.py
@@ -82,21 +82,6 @@ def ensure_local_garden():
     garden.save()
 
 
-def ensure_trigger_migration():
-    """File Triggers didn't work all the time, they were removed in 3.14.
-    Remove any file trigger jobs."""
-
-    database = get_db()
-
-    for doc in database["job"].find():
-        try:
-            if doc["trigger_type"] == "file":
-                database["job"].delete_one({"_id": doc["_id"]})
-
-        except Exception:
-            logger.error(f"Error deleting file trigger {doc['_id']}")
-
-
 def ensure_v2_to_v3_model_migration():
     """Ensures that the Role model is flatten and Command model is an
     EmbeddedDocument
@@ -147,7 +132,6 @@ def ensure_model_migration():
     single function for easy management"""
 
     ensure_v2_to_v3_model_migration()
-    ensure_trigger_migration()
 
 
 def check_indexes(document_class):