+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 TBD
 
+- Added file trigger jobs to scheduler to monitor directory for file changes. If a watched file change event occurs, it will fire the
+specified request. File trigger can be customized using regex, recursively monitoring sub-directories, or by watching one or more file
+change events (i.e. create, modify, move, or delete).
 - Request auto refresh configurable through `ui.auto_refresh` config option.
 - Added additional matching for topic subscriptions to include full string, empty, None, and regular expressions.
 - Fixed Request system version to return actual system version instead of parsed version when using latest.